save/load data have more concrete code//Not ended

+++ Assets/Scripts/Tower/TowerUpgradesDatabase.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using UnityEngine;
 using System.Linq;
-using Newtonsoft.Json.Schema;
 
 public class TowerUpgradesDatabase 
 {
@@ -14,7 +13,7 @@ public class TowerUpgradesDatabase
         _allUpgrades = Resources.LoadAll<TowerUpgrade>("ScriptableObjects/TowerUpgrades");
     }
 
-    public IEnumerable<TowerUpgrade> GetTowerUpgrades(TowerType towerType,int currentLevel)
+    public IEnumerable<TowerUpgrade> GetTowerUpgrades(TowerType towerType,byte currentLevel)
     {
         var upgrades = from upgrade in _allUpgrades
                        where upgrade.GetTowerType() == towerType &&  upgrade.GetUpgradeLevel() <= currentLevel
